Time: 0 ms (100%), Space: 9.2 MB (73.11%) - LeetHub

+++ 0377-combination-sum-iv/0377-combination-sum-iv.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int combinationSum4(vector<int>& nums, int target) {
 
-        vector<int> dp(target + 1, 0);
+        vector<unsigned long long> dp(target + 1, 0);
         dp[0] = 1;
 
         for(int t = 1; t <= target; t++){
@@ -12,6 +12,7 @@ public:
                 }
             }
         }
-        return dp[target];
+
+        return (int)dp[target];
     }
 };
